change id type to size_t

+++ include/state-observation/tools/measurements-manager/IMU.hpp
@@ -15,7 +15,7 @@ namespace measurements
 struct IMU : public Sensor
 {
 public:
-  inline IMU(int id, std::string_view name) : Sensor(id, name) {}
+  inline IMU(size_t id, std::string_view name) : Sensor(id, name) {}
 
 public:
   Eigen::Vector3d gyroBias = Eigen::Vector3d::Zero();
